Escape-Sequence-new-line-tab-in-C
This is a simple C program that demonstrates the usage of the printf() function from the stdio.h library.

The int main() function is the entry point of the program. The program has four printf() statements that print out strings to the console.

The first printf() statement prints out the string "This is a" to the console, followed by a newline character '\n' which moves the cursor to the next line.

The second printf() statement prints out the string "C program" to the console, followed by a newline character '\n' which moves the cursor to the next line.

The third printf() statement prints out the string "This", followed by four newline characters '\n' and the string "is", then another newline character, the string "a", another newline character, the string "C", and finally a newline character followed by the string "program". This results in each string being printed on a new line.

The fourth printf() statement uses the tab escape sequence '\t' to separate the string "This", "is", "a", and "C program" with tabs. This will align each string to the next tab stop in the console.

Finally, the program returns 0 using the return statement, which signals to the operating system that the program has executed successfully.